+++ packages/react-instantsearch/rollup.config.js
@@ -45,8 +45,8 @@ const plugins = [
 ];
 
 const createConfiguration = ({ name, minify = false } = {}) => ({
-  input: 'src/index.umd.ts',
-  external: ['react', 'ai'],
+  input: 'src/index.ts',
+  external: ['react'],
   output: {
     file: `dist/umd/ReactInstantSearch${name}${minify ? '.min' : ''}.js`,
     name: `ReactInstantSearch${name}`,