+++ packages/create-instantsearch-app/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [7.5.36](https://github.com/algolia/instantsearch/compare/create-instantsearch-app@7.5.35...create-instantsearch-app@7.5.36) (2026-01-27)
+
+**Note:** Version bump only for package create-instantsearch-app
+
+
+
+
+
 ## [7.5.35](https://github.com/algolia/instantsearch/compare/create-instantsearch-app@7.5.34...create-instantsearch-app@7.5.35) (2025-12-30)
 
 **Note:** Version bump only for package create-instantsearch-app